microlane/models/lanenet2/lanenet2/requirements.txt
absl-py==1.4.0
aiocontextvars==0.2.2
astor==0.8.1
cached-property==1.5.2
contextvars==2.4
cycler==0.11.0
dataclasses==0.8
gast==0.2.2
google-pasta==0.2.0
grpcio==1.48.2
h5py==3.1.0
immutables==0.19
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
loguru==0.7.3
Markdown==3.3.7
matplotlib==3.3.4
numpy==1.19.5
opencv-contrib-python==4.5.5.64
opt-einsum==3.3.0
Pillow==8.4.0
protobuf==3.19.6
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyYAML==6.0.1
scikit-learn==0.24.2
scipy==1.5.4
six==1.17.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
threadpoolctl==3.1.0
tqdm==4.64.1
typing_extensions==4.1.1
Werkzeug==2.0.3
wrapt==1.16.0
zipp==3.6.0